refactor(reputation): add a folder for each validator
re #201
